[FFM-5307] - .NET SDK - Fix for prereq identifier + update JSON parser to use latest tests
What
When a feature depends on another prereq feature being true, the evaluation fails if the value and identifier are different values in the prereq. Update code to use latest ff-test-cases JSON format.
Some additional fixes were needed including Null reference exception and incorrect OR handling in groups when latest JSON was used.

Why
The checkPreRequisite() method incorrectly uses the value of the prereq instead of the identifier.

Testing
Tested locally with new ff-test-cases JSON + sanity checks with example program