(FFM-8128): Fixes handling of null targets
WHAT: When a user passes in a `null` as the Target value of an evaluation
request, the metrics processor received a NullPointerException trying to parse
the target and add it to the TargetData.

FIX: The fix is to allow the logging and evaluation logic to handle the null
value as opposed to trying to dereference the Target attributes. Also in the
prepareMessageBody function, to handle the Target Data parsing separately from
the evaluation metrics, so that evaluation counts still get sent in, even if
the Target data has a null value.