[FFM-11056] - Optimize IN clause rules (#109)
* [FFM-11056] - Optimize IN clause rules

What
Cache values into a `HashSet` when `IN` clause is used

Why
The API generated code is using a list for values, so when the .NET evaluator processes an `IN` clause it has to scan the entire list. If thousands of value entries are present this can slow down rules processing a lot as the list lookup is O(n). We should instead cache this into a `HashSet` to get O(1) lookup speed on flag evaluation.

Testing
Manual + testgrid