[FFM-11557] - Wrap log statements with if statements
What
Wrap some debug statements on the evaluation path with if statements so they don't get run when logging is disabled for that level.
Also wrapped some warning loggers too

Why
We're seeing debug logs creating noise on the heap when log level is set to something higher. We want to prevent unnecessary objects on the heap/work for the GC.

Testing
Load + Testgrid